% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/populartimes.R
\name{get_place_details}
\alias{get_place_details}
\title{Query Google Place Detail API}
\usage{
get_place_details(place_id, key = Sys.getenv("GOOGLE_KEY"))
}
\arguments{
\item{place_id}{Google place ID. See \href{https://developers.google.com/maps/documentation/javascript/examples/places-placeid-finder}{docs} for app to provide place ID.}

\item{key}{Google API key from https://console.cloud.google.com/. Ensure Places API is enabled.}
}
\description{
Query Google Place Detail API
}
