% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_places.R
\name{get_places}
\alias{get_places}
\title{Search for Google Places in a given area}
\usage{
get_places(
  location = NULL,
  radius = NULL,
  type = NULL,
  keyword = NULL,
  language = NULL,
  rankby = NULL,
  key = Sys.getenv("GOOGLE_KEY")
)
}
\arguments{
\item{location}{A numeric vector of length two with the format of \code{c(lat, long)} indicating the center of a circle to search.}

\item{radius}{The radius of the search circle in meters.}

\item{type}{A place type to limit the search to. See API \href{https://developers.google.com/places/web-service/supported_types}{docs} for valid place types. Only one type may be specified per search.}

\item{key}{Your Google Places API key.}
}
\description{
Search for Google Places in a given area
}
