% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/populartimes.R
\name{get_popular_times}
\alias{get_popular_times}
\title{Search for Google popular times}
\usage{
get_popular_times(place_id, key = Sys.getenv("GOOGLE_KEY"))
}
\arguments{
\item{place_id}{Google place ID. See \href{https://developers.google.com/maps/documentation/javascript/examples/places-placeid-finder}{docs} for app to provide place ID.}

\item{key}{Google API key from https://console.cloud.google.com/. Ensure Places API is enabled.}
}
\description{
Search for Google popular times
}
