% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/place_types.R
\docType{data}
\name{place_types}
\alias{place_types}
\title{Google Places Types}
\format{
A data frame containing one column \code{place_types} with 96 values that can be used in the \code{type} argument of \code{get_places()} and \code{search_pop_times()}.
}
\source{
\url{https://developers.google.com/places/web-service/supported_types}
}
\usage{
place_types
}
\description{
Google Places Types
}
\keyword{datasets}
