% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poptimes_from_address.R
\name{poptimes_from_address}
\alias{poptimes_from_address}
\title{Get the popular times given a place name and address}
\usage{
poptimes_from_address(place_name, address)
}
\arguments{
\item{place_name}{The name of the place.}

\item{address}{A string containing the address of a place. Roughly as "\if{html}{\out{<street address>}}, \if{html}{\out{<city>}}, \if{html}{\out{<region>}}, \if{html}{\out{<postal code>}}, \if{html}{\out{<country>}}". This field is flexible.}
}
\description{
Search Google for popular times for a given point of interest's name and address. Google is smart, so the address component is not exceptionally strict. But do your best to format it!
}
\examples{
poptimes_from_address("Red Arrow", "61 Lowell Street, Manchester, NH, 03101, US")
poptimes_from_address("McDonald", "Helmstraat 16, 6211 TA Maastricht, Netherlands")
}
