% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_pop_times.R
\name{search_pop_times}
\alias{search_pop_times}
\title{Search an area for popular times.}
\usage{
search_pop_times(
  sw,
  ne,
  radius = 180,
  type = NULL,
  key = Sys.getenv("GOOGLE_KEY"),
  ...
)
}
\arguments{
\item{sw}{A numeric vector of length two with the format of c(lat, long) indicating the southwest point of a bounding box. Must be in decimal degrees and EPSG 4326.}

\item{ne}{A numeric vector of length two with the format of c(lat, long) indicating northeast point of a bounding box. Must be in decimal degrees and EPSG 4326.}

\item{radius}{The search radius of each grid element in meters.}

\item{type}{A place type to limit the search to. See API \href{https://developers.google.com/places/web-service/supported_types}{docs} for valid place types. Only one type may be specified per search.}

\item{key}{Your Google Places API key.}
}
\description{
Search an area for popular times.
}
\details{
The Google Places API searches return a maximum of 60 responses regardless of a search's radius. To accommodate this limitation of the API, a search grid is created with a number of smaller circles with a defined radius via the \code{radius} argument. Each cell in the grid uses a single API call. A smaller radius should be used to accommodate denser geographies.
The search grid used is saved the the attribute \code{search_grid} and can be accessed \code{attr(x, "search_grid")}.
}
\examples{
sw <- c(42.988690, -71.465834)
ne <- c(42.995119, -71.455745)
manch_bars <- search_pop_times(sw, ne, radius = 200, type = "bar")
}
